+++ modules/apps/foundation/friendly-url/friendly-url-service/src/main/java/com/liferay/friendly/url/model/impl/FriendlyURLEntryImpl.java
@@ -16,6 +16,13 @@ package com.liferay.friendly.url.model.impl;
 
 import aQute.bnd.annotation.ProviderType;
 
+import com.liferay.friendly.url.model.FriendlyURLEntryLocalization;
+import com.liferay.friendly.url.service.FriendlyURLEntryLocalizationLocalServiceUtil;
+import com.liferay.portal.kernel.util.LocaleUtil;
+import com.liferay.portal.kernel.util.StringPool;
+
+import java.util.Locale;
+
 /**
  * @author Pavel Savinov
  */
@@ -23,6 +30,34 @@ import aQute.bnd.annotation.ProviderType;
 public class FriendlyURLEntryImpl extends FriendlyURLEntryBaseImpl {
 
 	public FriendlyURLEntryImpl() {
+	}
+
+	@Override
+	public String getUrlTitle(Locale locale) {
+		String languageId = LocaleUtil.toLanguageId(locale);
+
+		FriendlyURLEntryLocalization friendlyURLEntryLocalization =
+			FriendlyURLEntryLocalizationLocalServiceUtil.
+				fetchFriendlyURLEntryLocalization(this, languageId);
+
+		if (friendlyURLEntryLocalization != null) {
+			return friendlyURLEntryLocalization.getUrlTitle();
+		}
+
+		return StringPool.BLANK;
+	}
+
+	@Override
+	public boolean isLocalized() {
+		int count =
+			FriendlyURLEntryLocalizationLocalServiceUtil.
+				getFriendlyURLEntryLocalizationsCount(this);
+
+		if (count > 0) {
+			return true;
+		}
+
+		return false;
 	}
 
 }