+++ modules/sdk/gradle-plugins-defaults/src/main/java/com/liferay/gradle/plugins/defaults/internal/LiferayRelengPlugin.java
@@ -144,7 +144,7 @@ public class LiferayRelengPlugin implements Plugin<Project> {
 				project, recordArtifactTask, cleanArtifactsPublishCommandsTask,
 				mergeArtifactsPublishCommandsTask);
 
-		mergeArtifactsPublishCommandsTask.dependsOn(
+		mergeArtifactsPublishCommandsTask.mustRunAfter(
 			writeArtifactPublishCommandsTask);
 
 		_addTaskPrintStaleArtifact(project, recordArtifactTask);
@@ -227,22 +227,28 @@ public class LiferayRelengPlugin implements Plugin<Project> {
 				public void execute(Task task) {
 					MergeFilesTask mergeFilesTask = (MergeFilesTask)task;
 
+					Logger logger = mergeFilesTask.getLogger();
+
 					File file = mergeFilesTask.getOutputFile();
 
 					boolean success = file.setExecutable(true);
 
 					if (!success) {
-						Logger logger = mergeFilesTask.getLogger();
-
 						logger.error(
 							"Unable to set the owner's execute permission " +
-								"for " + file);
+								"for {}",
+							file);
+					}
+
+					if (logger.isLifecycleEnabled()) {
+						logger.lifecycle(
+							"Artifacts publish commands written in {}.", file);
 					}
 				}
 
 			});
 
-		mergeFilesTask.setDescription("Merges the artifacts publish commands");
+		mergeFilesTask.setDescription("Merges the artifacts publish commands.");
 		mergeFilesTask.setHeader(
 			"#!/bin/bash" + System.lineSeparator() + System.lineSeparator() +
 				"set -e" + System.lineSeparator());